ここから生徒の学年情報をどうにかしないといけない

+++ demo/src/main/java/com/example/demo/entity/Student.java
@@ -16,7 +16,7 @@ public class Student {
     private Integer id;
 
     @Column(name="code")
-    private Integer code;
+    private Long code;
 
     @Column(name="name")
     private String name;
